Reorganise Api project structure
