+++ QuestionServiceWebApi/Interfaces/IQuestionRepository.cs
@@ -1,3 +1,5 @@
+using QuestionServiceWebApi.Entities;
+
 namespace QuestionServiceWebApi.Interfaces
 {
     public interface IQuestionRepository
