+++ QuestionServiceWebApi/QuestionServiceWebApi.csproj
@@ -137,8 +137,8 @@
     </Compile>
     <Compile Include="Interfaces\IQuestionRepository.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Questionnaire.cs" />
-    <Compile Include="QuestionRepository.cs" />
+    <Compile Include="Entities\Questionnaire.cs" />
+    <Compile Include="Repositories\QuestionRepository.cs" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="App_Start\FilterConfig.cs" />
